+++ src/connectionMatrix/node.cpp
@@ -225,9 +225,9 @@ OfflineOutputStreamNode::OfflineOutputStreamNode() noexcept
 /* ************************************************************ */
 /* EntityNode                                                   */
 /* ************************************************************ */
-EntityNode* EntityNode::create(la::avdecc::UniqueIdentifier const& entityID, bool const isMilan, bool const isRegisteredUnsol, bool const areUnsolSupported) noexcept
+EntityNode* EntityNode::create(la::avdecc::UniqueIdentifier const& entityID, bool const isMilan, la::avdecc::entity::model::MilanVersion const& milanCompatibleVersion, bool const isRegisteredUnsol, bool const areUnsolSupported) noexcept
 {
-	return new EntityNode{ entityID, isMilan, isRegisteredUnsol, areUnsolSupported };
+	return new EntityNode{ entityID, isMilan, milanCompatibleVersion, isRegisteredUnsol, areUnsolSupported };
 }
 
 void EntityNode::accept(la::avdecc::entity::model::AvbInterfaceIndex const avbInterfaceIndex, AvbInterfaceIndexVisitor const& visitor) const noexcept
@@ -246,9 +246,10 @@ void EntityNode::accept(la::avdecc::entity::model::AvbInterfaceIndex const avbIn
 		});
 }
 
-EntityNode::EntityNode(la::avdecc::UniqueIdentifier const& entityID, bool const isMilan, bool const isRegisteredUnsol, bool const areUnsolSupported) noexcept
+EntityNode::EntityNode(la::avdecc::UniqueIdentifier const& entityID, bool const isMilan, la::avdecc::entity::model::MilanVersion const& milanCompatibleVersion, bool const isRegisteredUnsol, bool const areUnsolSupported) noexcept
 	: Node{ Type::Entity, entityID, nullptr }
 	, _isMilan{ isMilan }
+	, _milanCompatibleVersion{ milanCompatibleVersion }
 	, _isRegisteredUnsol{ isRegisteredUnsol }
 	, _areUnsolSupported{ areUnsolSupported }
 {
@@ -282,6 +283,11 @@ void EntityNode::setOutputAudioMappings(la::avdecc::entity::model::StreamPortInd
 bool EntityNode::isMilan() const noexcept
 {
 	return _isMilan;
+}
+
+la::avdecc::entity::model::MilanVersion const& EntityNode::milanCompatibleVersion() const noexcept
+{
+	return _milanCompatibleVersion;
 }
 
 bool EntityNode::isRegisteredUnsol() const noexcept
@@ -596,35 +602,62 @@ void StreamNode::computeIsStreaming() noexcept
 /* ************************************************************ */
 /* ChannelNode                                                  */
 /* ************************************************************ */
-ChannelNode* ChannelNode::createOutputNode(EntityNode& parent, avdecc::ChannelIdentification const& channelIdentification) noexcept
+ChannelNode* ChannelNode::createOutputNode(EntityNode& parent, la::avdecc::controller::model::ClusterIdentification const& clusterIdentification, avdecc::ChannelIdentification const& channelIdentification_old) noexcept
 {
-	return new ChannelNode{ Type::OutputChannel, parent, channelIdentification };
+	return new ChannelNode{ Type::OutputChannel, parent, clusterIdentification, channelIdentification_old };
 }
 
-ChannelNode* ChannelNode::createInputNode(EntityNode& parent, avdecc::ChannelIdentification const& channelIdentification) noexcept
+ChannelNode* ChannelNode::createInputNode(EntityNode& parent, la::avdecc::controller::model::ClusterIdentification const& clusterIdentification, avdecc::ChannelIdentification const& channelIdentification_old, la::avdecc::controller::model::ChannelIdentification const& channelIdentification) noexcept
 {
-	return new ChannelNode{ Type::InputChannel, parent, channelIdentification };
+	return new ChannelNode{ Type::InputChannel, parent, clusterIdentification, channelIdentification_old, channelIdentification };
 }
 
-avdecc::ChannelIdentification const& ChannelNode::channelIdentification() const noexcept
+avdecc::ChannelIdentification const& ChannelNode::channelIdentification_old() const noexcept
+{
+	return _channelIdentification_old;
+}
+
+la::avdecc::controller::model::ClusterIdentification const& ChannelNode::clusterIdentification() const noexcept
+{
+	return _clusterIdentification;
+}
+
+la::avdecc::controller::model::ChannelIdentification const& ChannelNode::channelIdentification() const noexcept
 {
 	return _channelIdentification;
 }
 
-la::avdecc::entity::model::ClusterIndex ChannelNode::clusterIndex() const noexcept
+bool ChannelNode::hasPrimaryMapping() const noexcept
 {
-	return _channelIdentification.clusterIndex;
+	return _hasPrimaryMapping;
 }
 
-std::uint16_t ChannelNode::channelIndex() const noexcept
+bool ChannelNode::hasSecondaryMapping() const noexcept
 {
-	return _channelIdentification.clusterChannel;
+	return _hasSecondaryMapping;
 }
 
-ChannelNode::ChannelNode(Type const type, Node& parent, avdecc::ChannelIdentification const& channelIdentification) noexcept
+ChannelNode::ChannelNode(Type const type, Node& parent, la::avdecc::controller::model::ClusterIdentification const& clusterIdentification, avdecc::ChannelIdentification const& channelIdentification_old, la::avdecc::controller::model::ChannelIdentification const& channelIdentification) noexcept
 	: Node{ type, parent.entityID(), &parent }
+	, _clusterIdentification{ clusterIdentification }
+	, _channelIdentification_old{ channelIdentification_old }
 	, _channelIdentification{ channelIdentification }
 {
+}
+
+void ChannelNode::setChannelIdentification(la::avdecc::controller::model::ChannelIdentification const& channelIdentification) noexcept
+{
+	_channelIdentification = channelIdentification;
+}
+
+void ChannelNode::setHasPrimaryMapping(bool const hasMapping) noexcept
+{
+	_hasPrimaryMapping = hasMapping;
+}
+
+void ChannelNode::setHasSecondaryMapping(bool const hasMapping) noexcept
+{
+	_hasSecondaryMapping = hasMapping;
 }
 
 /* ************************************************************ */